float

+++ eigvals.py
@@ -3,7 +3,7 @@ import numpy as np
 n = int(input())
 A=[]
 for i in range(n):
-    A.append(list(map(int, input().split())))
+    A.append(list(map(float, input().split())))
 
 eigs = np.linalg.eigvals(A)
 summ_of_eigs = np.sum(eigs)
